move VCalendar component into lib/calagator.
